production/requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic[email]==2.7.1
python-dotenv==1.0.1

# AI / Agents
openai-agents==0.0.8
anthropic==0.28.0
litellm==1.40.0
mcp==1.0.0

# Database
asyncpg==0.29.0
pgvector==0.3.2

# Kafka
aiokafka==0.11.0

# Gmail
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-oauthlib==1.2.0

# WhatsApp / Twilio
twilio==9.1.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
locust==2.28.0

# Utilities
python-multipart==0.0.9
